My exam/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.2.0
auto-py-to-exe==2.29.2
bottle==0.12.23
bottle-websocket==0.2.9
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
Eel==0.14.0
et-xmlfile==1.1.0
filelock==3.10.7
Flask==2.2.3
frozenlist==1.3.3
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.16.2
MarkupSafe==2.1.2
multidict==6.0.4
openpyxl==3.1.2
platformdirs==3.2.0
psycopg2-binary==2.9.5
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2022.15
pynput==1.7.6
pyobjc-core==9.0.1
pyobjc-framework-ApplicationServices==9.0.1
pyobjc-framework-Cocoa==9.0.1
pyobjc-framework-Quartz==9.0.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.0
PyQt6==6.4.2
PyQt6-Qt6==6.4.3
PyQt6-sip==13.4.1
requests==2.28.2
six==1.16.0
urllib3==1.26.14
virtualenv==20.21.0
Werkzeug==2.2.3
whichcraft==0.6.1
yarl==1.8.2
zope.event==4.6
zope.interface==5.5.2
